refactor: RecordForm, RecordList, RecordItem 컴포넌트 폴더 구조 분리 #5
